@@ -5,7 +5,7 @@ buildscript {
         google()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:3.3.1'
     }
 }
 
@@ -20,25 +20,35 @@ repositories {
 }
 
 dependencies {
-    implementation 'com.android.support:design:27.1.1'
-    implementation 'com.android.support:recyclerview-v7:27.1.1'
-    implementation 'com.android.support:cardview-v7:27.1.1'
-    implementation 'com.android.support:appcompat-v7:27.1.1'
-    implementation 'com.github.bumptech.glide:glide:3.7.0' //4.8.0' //3.7.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    //implementation 'androidx.annotation:annotation:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
     implementation 'org.jsoup:jsoup:1.11.3'
+
+    implementation 'com.github.bumptech.glide:glide:3.7.0'  // 4.8.0' //3.7.0'
+    //annotationProcessor 'com.github.bumptech.glide:compiler:4.8.0'
+
+    implementation 'net.dankito.readability4j:readability4j:1.0.2'
+
+    def okhttp_version = '3.12.1'
+    implementation "com.squareup.okhttp3:okhttp:$okhttp_version"
+    implementation "com.squareup.okhttp3:okhttp-urlconnection:$okhttp_version"
+
 }
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     buildToolsVersion "28.0.3"
 
     defaultConfig {
         applicationId "de.bernd.shandschuh.sparserss"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         compileOptions {
-            sourceCompatibility JavaVersion.VERSION_1_7
-            targetCompatibility JavaVersion.VERSION_1_7
+            sourceCompatibility JavaVersion.VERSION_1_8
+            targetCompatibility JavaVersion.VERSION_1_8
         }
     }
 
